pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools-scm>=8", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pyo"
version = "0.0.0"
description = "Numerical piano with manual/automatic playback and didactic overlays"
readme = "README.md"
authors = [{name = "Thomas COUPPEY"}]
requires-python = ">=3.10"
dependencies = [
    "mido>=1.3",
    "numpy>=1.26",
    "pandas>=2.3",
    "pydantic>=2.0",
    "pygame-ce>=2.5",
    "screeninfo>=0.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "black>=24.0",
    "ruff>=0.3",
    "mypy>=1.9",
]
examples = [
    "jupyterlab>=4.1",
]

[project.scripts]
pyo = "pyo.cli:app"

[tool.setuptools]
package-dir = {"" = "pyo"}

[tool.black]
target-version = ["py310"]
line-length = 100

[tool.ruff]
target-version = "py310"
line-length = 100
lint.select = ["E", "F", "I", "UP", "N", "B", "A"]
lint.ignore = ["E402", "I001"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q"
testpaths = ["tests"]
